requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.7.0
django-graphql-jwt==0.4.0
graphene==3.4.3
graphene-django==3.2.3
graphene-file-upload==1.3.0
graphql-core==3.2.6
graphql-relay==3.2.0
pillow==11.2.1
promise==2.3
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.0
